refactor(constants): 统一常量文件命名为 kebab-case 格式
- 为什么改：项目中的 handlers、services 等目录已统一使用 kebab-case 命名规范，constants 目录仍使用
PascalCase，为保持项目命名规范一致性，需要进行统一
- 改了什么：
  1. 重命名 6 个常量文件：ApiConstants → api.constants、CacheConstants → cache.constants、EventConstants →
event.constants、HttpConstants → http.constants、McpConstants → mcp.constants、TimeoutConstants → timeout.constants
  2. 更新 constants/index.ts 中的 6 个导出语句
  3. 更新 mcp-tool-log.handler.ts 中的导入路径
  4. 更新 mcp-tool-log.handler.test.ts 中的导入路径
- 影响范围：仅涉及文件重命名和导入路径更新，无逻辑变更，无破坏性变更
- 验证方式：
  1. 所有测试通过（965 个测试用例全部通过）
  2. TypeScript 类型检查通过
  3. Biome lint 检查通过
  4. 拼写检查和重复代码检查通过
