+++ apps/backend/constants/event.constants.ts
