+++ apps/backend/constants/index.ts
@@ -14,19 +14,19 @@
  */
 
 // API 相关常量
-export * from "./ApiConstants.js";
+export * from "./api.constants.js";
 
 // HTTP 协议常量
-export * from "./HttpConstants.js";
+export * from "./http.constants.js";
 
 // MCP 协议常量
-export * from "./McpConstants.js";
+export * from "./mcp.constants.js";
 
 // 事件名称常量
-export * from "./EventConstants.js";
+export * from "./event.constants.js";
 
 // 超时和延迟常量
-export * from "./TimeoutConstants.js";
+export * from "./timeout.constants.js";
 
 // 缓存相关常量
-export * from "./CacheConstants.js";
+export * from "./cache.constants.js";
